Removed schema factories that aren't currently useful
